configure: support --enable-xen and --enable-xenctrl arguments
Previously `opam` would install a depopt and then recompile this package
expecting there to be some side-effect. If there is a problem with the
build environment (e.g. a missing C header) then the side-effect would
be skipped, resulting in downstream `vchan.lwt` not found errors.

The new `--enable-xenctrl` and `--enable-xen` arguments allow `opam` to
force enable the particular options, triggering an early failure if
the build environment is broken.

Fixes #72

Signed-off-by: David Scott <dave@recoil.org>